refactor(aztec-node): extract block-parameter helpers as free functions
Move normalizeBlockParameter, isBlockTag, isCheckpointTag, and resolveCheckpointParameter out of
`AztecNodeService` into `src/modules/block_parameter.ts` as pure free functions (resolveCheckpointParameter
takes the block source as an argument for its tip lookups).

`NodeWorldStateQueries` now imports `normalizeBlockParameter` directly instead of receiving it as an
injected dependency, and `AztecNodeService` calls the free functions from its remaining block/checkpoint
read methods.
